OrganizationConnect/requirements.txt
 
Django==3.2.12
django-active-link 
django-ckeditor 
django-countries 
django-environ 
django-js-asset 
django-taggit 
 
 
 
 

asgiref
 
django-cleanup
django-crispy-forms
django-solo
django-tinymce4-lite
jsmin
 
Pillow
python-dotenv
pytz
sqlparse
crispy-bootstrap4